+++ src/server/aria2c_stub.go
@@ -1,0 +1,7 @@
+//go:build !darwin
+
+package main
+
+func darwinAria2cExtraCandidates() []string { return nil }
+
+func ensureAria2cDarwinAtStartup() error { return nil }